refactor views in recipes into recipes_view and single_recipe_view
